Move lunaria config to main package
