+++ lunaria/renderer.config.ts
@@ -1,0 +1,10 @@
+import { defineRendererConfig } from '@lunariajs/core';
+
+import { TitleParagraph, TranslationNeedsReview } from './components';
+
+export default defineRendererConfig({
+  slots: {
+    afterTitle: TitleParagraph,
+    afterStatusByLocale: TranslationNeedsReview,
+  },
+});
