+++ package.json
@@ -19,7 +19,7 @@
     "build:cli": "pnpm --filter cli-generator run build",
     "build:astro": "astro build",
     "build:i18n": "pnpm --filter lunaria run build",
-    "build": "pnpm dev:setup && pnpm build:references && pnpm build:config && pnpm build:cli && pnpm build:releases && pnpm build:astro && pnpm build:i18n",
+    "build": "pnpm dev:setup && pnpm build:references && pnpm build:config && pnpm build:cli && pnpm build:releases && pnpm build:astro",
     "preview": "astro preview",
     "lunaria:build": "pnpm --filter lunaria run build",
     "lunaria:preview": "pnpm --filter lunaria run preview"
@@ -29,6 +29,8 @@
     "@astrojs/rss": "^4.0.7",
     "@astrojs/starlight": "0.29.0",
     "@lorenzo_lewis/starlight-utils": "^0.2.0",
+    "@lunariajs/core": "^0.1.1",
+    "@lunariajs/starlight": "^0.1.1",
     "@types/json-schema": "^7.0.15",
     "astro": "^4.15.7",
     "astro-d2": "^0.6.0",
