--- packages/lunaria/lunaria.config.json
@@ -1,50 +1,0 @@
-{
-  "$schema": "./node_modules/@lunariajs/core/config.schema.json",
-  "repository": {
-    "name": "tauri-apps/tauri-docs",
-    "branch": "v2",
-    "rootDir": "src/content/docs",
-    "hosting": "github"
-  },
-  "files": [
-    {
-      "location": "../../src/content/docs/**/*.{md,mdx}",
-      "pattern": "../../src/content/docs/@lang/@path",
-      "ignore": [
-        "node_modules",
-        "../../src/content/docs/reference/acl/**/*.{md,mdx}",
-        "../../src/content/docs/reference/javascript/**/*.{md,mdx}",
-        "../../src/content/docs/reference/{config,cli}.{md,mdx}",
-        "../../src/content/docs/release"
-      ],
-      "type": "universal"
-    }
-  ],
-  "outDir": "../../dist/contribute/translate-status",
-  "localizableProperty": "i18nReady",
-  "renderer": "./renderer.config.ts",
-  "ignoreKeywords": ["[en-only]", "[typo]", "[i18nReady]", "[i18nIgnore]", "[format]", "[ci]"],
-  "defaultLocale": {
-    "label": "English",
-    "lang": "en"
-  },
-  "locales": [
-    {
-      "label": "Français",
-      "lang": "fr"
-    },
-    {
-      "label": "Español",
-      "lang": "es"
-    },
-    {
-      "label": "简体中文",
-      "lang": "zh-cn"
-    }
-  ],
-  "dashboard": {
-    "title": "Tauri Docs Translation Status",
-    "basesToHide": ["../../src/content/docs/"],
-    "customCss": ["./style.css"]
-  }
-}
