--- packages/lunaria/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "lunaria",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build": "lunaria build",
-    "preview": "lunaria preview"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@lunariajs/core": "^0.1.1",
-    "node-fetch": "^3.3.2",
-    "p-retry": "^6.2.0"
-  }
-}
