+++ pnpm-workspace.yaml
@@ -6,4 +6,3 @@ packages:
   - 'packages/tauri-typedoc-theme'
   - 'packages/releases-generator'
   - 'packages/compatibility-table'
-  - 'packages/lunaria'
